[WIP] Vectors.h renamed to Renderer.h
